LayoutTests/remote-layer-tree/no-layer-changes-from-raf-expected.txt

PASS initialCountOfLayerChangingUpdates === finalCountOfLayerChangingUpdates is true
PASS successfullyParsed is true

TEST COMPLETE
